backend/app/api/v1/endpoints/tickets.py
from fastapi import APIRouter, Depends, HTTPException, BackgroundTasks
from typing import List, Optional
from pydantic import BaseModel
from datetime import datetime
from enum import Enum
from opentelemetry import trace

router = APIRouter()
tracer = trace.get_tracer(__name__)

class TicketPriority(str, Enum):
    LOW = "low"
    MEDIUM = "medium"
    HIGH = "high"
    URGENT = "urgent"

class TicketStatus(str, Enum):
    NEW = "new"
    TRIAGED = "triaged"
    ASSIGNED = "assigned"
    IN_PROGRESS = "in_progress"
    RESOLVED = "resolved"
    CLOSED = "closed"

class TicketType(str, Enum):
    MAINTENANCE = "maintenance"
    CONCIERGE = "concierge"
    ADMINISTRATIVE = "administrative"

class TicketBase(BaseModel):
    title: str
    description: str
    property_id: str
    unit_id: Optional[str] = None
    tenant_id: str
    channel: str  # SMS, WhatsApp, Email, Voice
    language: str  # FR, EN, DE

class TicketCreate(TicketBase):
    pass

class Ticket(TicketBase):
    id: str
    created_at: datetime
    updated_at: datetime
    status: TicketStatus
    priority: TicketPriority
    type: TicketType
    assigned_to: Optional[str] = None
    resolution_time: Optional[datetime] = None

    class Config:
        from_attributes = True

@router.post("/", response_model=Ticket)
async def create_ticket(
    ticket: TicketCreate,
    background_tasks: BackgroundTasks
):
    """
    Create a new ticket and trigger AI triage in the background.
    """
    with tracer.start_as_current_span("create_ticket") as span:
        span.set_attribute("ticket.channel", ticket.channel)
        span.set_attribute("ticket.language", ticket.language)
        
        # TODO: Implement ticket creation in database
        # TODO: Implement AI triage in background
        background_tasks.add_task(perform_ai_triage, ticket)
        
        return {
            "id": "temp-id",
            "created_at": datetime.utcnow(),
            "updated_at": datetime.utcnow(),
            "status": TicketStatus.NEW,
            "priority": TicketPriority.MEDIUM,
            "type": TicketType.MAINTENANCE,
            **ticket.dict()
        }

@router.get("/{ticket_id}", response_model=Ticket)
async def get_ticket(ticket_id: str):
    """
    Get a specific ticket by ID.
    """
    with tracer.start_as_current_span("get_ticket") as span:
        span.set_attribute("ticket.id", ticket_id)
        # TODO: Implement ticket retrieval from database
        raise HTTPException(status_code=404, detail="Ticket not found")

@router.get("/", response_model=List[Ticket])
async def list_tickets(
    status: Optional[TicketStatus] = None,
    priority: Optional[TicketPriority] = None,
    type: Optional[TicketType] = None,
    property_id: Optional[str] = None,
    skip: int = 0,
    limit: int = 100
):
    """
    List tickets with optional filtering.
    """
    with tracer.start_as_current_span("list_tickets") as span:
        # TODO: Implement ticket listing from database
        return []

@router.put("/{ticket_id}", response_model=Ticket)
async def update_ticket(
    ticket_id: str,
    ticket_update: TicketCreate
):
    """
    Update a ticket.
    """
    with tracer.start_as_current_span("update_ticket") as span:
        span.set_attribute("ticket.id", ticket_id)
        # TODO: Implement ticket update in database
        raise HTTPException(status_code=404, detail="Ticket not found")

async def perform_ai_triage(ticket: TicketCreate):
    """
    Perform AI-based triage on a ticket to determine:
    1. Ticket type (maintenance, concierge, administrative)
    2. Priority level
    3. Initial routing
    """
    with tracer.start_as_current_span("perform_ai_triage") as span:
        # TODO: Implement AI triage logic
        # 1. Use transformers for classification
        # 2. Apply business rules
        # 3. Update ticket in database
        # 4. Trigger notifications if needed
        pass 