backend/requirements.txt
# API Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Message Queue
aio-pika==9.3.0
kafka-python==2.0.2

# AI/ML
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
prophet==1.1.4

# Cloud Storage
boto3==1.29.3
python-dotenv==1.0.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-oauth2==1.0.1

# Monitoring
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
aiohttp==3.9.1 